Apply small refactoring propositions
