+++ app/src/androidTest/java/com/github/leuludyha/ibdb/presentation/components/WorkListTest.kt
@@ -69,7 +69,7 @@ class WorkListTest {
             subjects = flowOf(listOf("Dog", "Murder Mystery"))
         )
 
-        // TODO how to passe lazyPagingItems?
+        // TODO how to test lazyPagingItems?
         /*composeTestRule.setContent {
             WorkList(
                 orientation = Orientation.Vertical,
