+++ data/src/test/java/com/github/leuludyha/data/api/RawSearchTest.kt
@@ -9,7 +9,7 @@ import org.junit.Test
 class RawSearchTest: RequiringLibraryApiTest() {
 
     @Test
-    fun `Fields are properly gotten`() {
+    fun `Fields are properly accessed`() {
         val doc = RawDocument(
             coverId = 0,
             title = "Awesome book",
