+++ data/src/test/java/com/github/leuludyha/data/io/FileReader.kt
@@ -1,4 +1,4 @@
-package com.github.leuludyha.data
+package com.github.leuludyha.data.io
 
 import java.io.IOException
 import java.io.InputStreamReader
